BUG#29279941 : check instance print error for loopback address
There was an inconsistency between the behavior of
dba.checkInstanceConfiguration() and the commands to add instances
to the cluster (createCluster() and addInstance()) regarding the
localhost and loopback address validation. In particular, a simple
error was printed by checkInstanceConfiguration() but the execution
of the operation continued showing that everything was "ok" at the
end, while an error is issued and the execution stopped for
createCluster() and addInstance().

In addition, attending to recent improvements (bug fix) that replaced
the use of the host specified in the connection by the value for
report_host or hostname of the instance, in order to be used
internally by InnoDB Cluster. It was decided that existing localhost
and loopback address validations are no longer needed and should be
removed. In particular, whatever address is specified for report_host
by the user, even if localhost or the loopback address (127.0.0.1),
should be allowed, since it was explicitly specified by the user to
use it.

This patch fixes the issue by removing all validation for the use of
localhost or host that resolves to the loopback address (127.0.0.1),
no longer issuing an error in those situations.

Change-Id: I9cae9a3a0d2f413e78db17602eb3b31c45324d78