UT Fix: Tests randomly failing due to clashes of ports
Some tests are using hardcoded ports which lead to random failures
whenever those ports are not properly released. The tests must use the
port ranges defined by the test-framework.

This patch removes the use of hardcoded ports guaranteeing the tests are
still valid.

On top of that, one test was randomly failing due to a missing call to
configureLocalInstance().

Change-Id: Iae78ec8f0bae6df9f8765b2f2d6acc3201cdff94