BUG#28393119 UNABLE TO GET HELP ON CONNECTION DATA
When a search query is done through "\? <pattern>" it is possible that
multiple topics are found matching the used pattern. On those cases
the topics are just listed so the user can do a specific search using
a fully qualified ID.

Two reasons cause the issue to happen:
- General topics do NOT have a fully qualified id, the topic title is
  the only way to match the topic.
- When a session is established, the help system will also look at the
  database for help and many other topics may be added to the result.

The topic "Connection" used to describe the way a connection is created
in the shell, happens to be a keyword that gives multiple matches when
a session is established, making it impossible to get the shell help for
it when a session is established.

This patch introduces an improvement on the help handling when there are
multiple matches:

- If exactly one of the found topics matches the given keyword (case
  sensitive) it is selected as the exact match.
- If no matches were found but one of the topics matches the given keyword
  (case insensitive) it is selected as the exact match.

If an exact match is found then it's help data will be printed and the rest
of the found topics will be listed in a SEE ALSO section.

If no exact match is found then all the topics will be listed.

Change-Id: I19c278ad5058ae67a3c617a808cadf2fc8721e33