BUG#29308037: NOT DELETE MYSQL_INNODB_CLUSTER_RXXX@'LOCALHOST'
Whenever dba.createCluster() fails and a rollback is performed to remove the
created replication (recovery) users, the account created at 'localhost'
and/or the 'ipWhitelist' corresponding addresses are not being removed.

This patch ensures that the replication accounts are removed whenever a
rollback at createCluster() is performed.

Added regression test to cover the scenario.

Based on a code contribution from Bin Hong.

Change-Id: If06a9d363e15fdde282c85d903492c607f01810c