BUG#29018457 : check instance does not validate account host
The dba.checkInstanceConfiguration() does not validate host
restrictions for the used account. In particular, an error should be
issued if the used user account is only able to connect through
'localhost'.

NOTE: The user account used for cluster administration must be able
to connect to all instance in the cluster (remotely).

This patch fixes the issue by adding the check for host restrictions
performed in configureInstance() to checkInstanceConfiguration().
The behavior of this check was also modified (improved) to print an
error instead of a warning/information message if some restriction is
found (e.g., can only connect from localhost), stopping the execution
of configureInstance in non-interactive mode.

In addition, an unused (related) function was removed from the Dba
interactive layer.

Tests were properly update to consider this change and new tests were
added for regression.

Change-Id: I4f23e32cbc53bad4d31a83172b76319082f74069