WL#12634 OCI: Add OCI Python APIs to MySQL Shell
Adds support for the OCI by:
- Bundling the Python OCI SDK.
- Introducing an OCI Setup wizard to configure different OCI profiles.

The wizard is accessible through the new shell.configureOci([profile])
function as well as through a new --oci[=profile] command line argument.

All the functionality is enabled with a WITH_OCI flag which should point
to a folder containing the OCI SDK as well as it's dependencies.

The key components of this WL include:

- A base Wizard class that allows creating interactive wizards in a
structured way.
- An Oci_setup wizard.
- SSL utilities for the handling of Key files.
- The bundled OCI SDK

Change-Id: I0702d3889659f5e1dc419b31a1561480fe74f042