BUG#29265869 : reboot cluster overrides some persisted settings
The dba.rebootClusterFromCompleteOutage() operation does not preserve
the existing Group Replication (GR) configurations previously set for
the instances. In particular, the GR local address and exit state
action values are changed.

This patch fixes this issue by reading the Group Replication
settings (that might be changed during the rejoin process) of each
instance, passing those values as input of the operation to bootstrap
and rejoin the instances to the cluster.

A new test was added for regression.

Change-Id: I2508b49ef7fd384571acdb6d6ce40ea36cd915ae