BUG#29356599 FAILOVERCONSISTENCY PURPOSE AND BEHAVIOR ERRONEOUS AND CONFUSING
The 'failoverConsistency' option name is not consistent with the target group
replication system variable name. In addition, the allowed values are not
documented which leads to confusion regarding what should be expected when the
option is assigned a sppecific value.

This patch deprecates this option in favor of a 'consistency' option, also
documents all the possible values as well as the expected results for each
of them.

This change impacts the following APIs
- dba.createCluster(...)
- cluster.setOption(...)

Change-Id: Id7033d0739606db578032c7457716ece92e8fbff