BUG#25686324 SHELL DOES NOT RECOGNIZE -F (EXECUTE FILE)
When Shell was launched with both --execute and --file options, it only
executed the given command, ignoring the given file.

Running Shell with both --execute and --file options is now disallowed,
resulting in an error. The output of --help command line option was
updated, both commands are marked as mutually exclusive.

Note: running Shell with both command line options in reverse order
(--file being the first one) does not result in an error, as all
options after --file are used as arguments of the processed file.

Change-Id: I2ac849d414639672a40264497742e90df8e6ae79