BUG#29305551 : ADMINAPI FAILS TO DETECT INSTANCE IS RUNNING ASYNCHRONOUS REPLICATION
Using an instance for InnoDB cluster, to add it or rejoin it to an existing
cluster, requires that asynchronous (master-slave) replication is not
configured and running.

dba.checkInstanceConfiguration() reports that a target instance which has the
asynchronous replication set up and running as valid for InnoDB cluster usage.
As a consequence, the AdminAPI fails to detect that an instance has
asynchronous replication running and both <Cluster.>addInstance() and
<Cluster.>rejoinInstance() fail with useless/unfriendly errors on this
scenario. On top of that, there's no information about the issue and how to
overcome it.

This patch extends dba.checkInstanceConfiguration() to verify if asynchronous
replication is configured and running on the target instance and print a
warning if that's the case. On top of that, the same check is used in
<Cluster.>addInstance() and <Cluster.>rejoinInstance() to terminate the
commands with an error if such scenario is verified.

Finally, the same check if used in dba.rebootClusterFromCompleteOutage()
whenever there are instances to be rejoined to the cluster.

Added new tests to cover all the possible scenarios that would hit this issue.

Change-Id: Ic2fbac58a46e35c1aa3d608e0d59b90638e103a2