WL#12011 : AdminAPI: Refactor dba.createCluster()
This patch refactors the dba.createCluster() command according to the
newly defined class structure for the AdminAPI, and it removes the dependency
from mysqlprovision (internal Python tool) to execute this operation (i.e.,
converting the start-replicaset command from mysqlprovision to C++). The main
behavior of the operation is maintained.

Existing tests already cover the performed changes.

In addition, it ensures BUG#28064729 and BUG#29361352 are fixed,
including specific tests for regression.

Change-Id: Iae934bb31de9c30f569a62e83cbab6f660bdf881