Misc Help Enhancements
This push introduces the following enhancements to the help system:
1) Adds member naming style resolution even for Doxygen generated help.
2) Adds support for <br> in shell help.
3) Replaces <code></code> for @code @endcode

Collateral change: detected that <code> was being used in a wrong way to
create a hardcoded formatting on the Command Line integration help. Have
reformatted that help to use standard formating.

Naming Style Resolution in Doxygen
----------------------------------
Resolution of member names has been supported in the shell help by simply
enclosing the member name as <<<camelCaseMemberName>>>.

However this did not work for Doxygen and we ended adding chunks of help
defined between DOXYGEN_JS/DOXYGEN_PY whenever we wanted to refer to another
member on the help.

Now this resolution also works for Doxygen as long as it is defined in any
of the REGISTER_*HELP* macros, this is, does not work if using <<<>>> directly
in doxygen comments.

Support for <br>
----------------
This is simply handy, it will be rendered as a new line character on the
shell help.

New Code Commands
-----------------
It was detected that <code></code> were not achieving the desired effect in
the Doxygen docs (which is creating a code snippet). For this reason we have
replaced it in the shell to use @code and @endcode commands instead.

Suggestion that whenever @code is required, the help is registered using the
new REGISTER_*HELP* macros that allow defining the entire help in a multiline
text definition, reason for this is because whatever spacing you use, it will
be rendered that way with no extra effort.

If not done that way you will have to give the desired formatting to the
text, i.e. adding new lines, using &nbsp, etc.

For examples about #1 look at the changes in mod_shell.cc
For examples about #3 the good way, look at the changes in
mod_mysqlx_collection.cc
For examples about #4 the bad way, look at the changes in mod_util.cc

Change-Id: I50e0812d86efa528abde8c690d96157c103ac66d