Fixed ReplicaSet::remove_replication_users() to avoid removing the wrong set of users:
Removing an instance from a cluster when the instance to be removed has
no user defined for channel group_replication_recovery in
mysql.slave_master_info results in dropping ALL users from mysql.user on the
remaining instances of the cluster.

In order to obtain the recovery account used by the instance to be
removed, the mysql.slave_master_info table is queried. However, when
that table is cleared out the query returns an empty string which is
then appended to the regex ".*" to remove the account from all the
cluster instances. Such regex results in ALL the accounts being removed.

This patch fixes the bug by validating if the query returns an empty
value and also by removing the usage of regex expressions to drop
accounts.

Added new regression tests.

Change-Id: I706be8a19309ef9db3add9e8c7279e4edb0427e2