WL#13051 Support for Shell Extensions
This worklog introduces support for shell extensions in the form
of plugins.

Plugins are read from the following folders:

<install>/share/mysqlsh/plugins
<shell user config path>/plugins

A plugin is a folder containing an initialization script (init.py/init.js)
and can be used to extend shell functionality, right now it supports:

- Definition of custom reports
- Definition of user objects

They are loaded at shell startup.

Side Changes
- Renamed existing Base_shell::load_plugins to Mysql_shell::load_startup_scripts.
- Added delayed printing logic into the Command_line_shell class.
- Updated paratraces to add sequence so every test is executed in an isolated
  environment when running in parallel.

Change-Id: Ia9526f6956a3f04516fae02d2379e6e0389ff713