BUG#29725222: sandboxes deployed via adminapi do not support restart
This patch adds support for the restart command on the sandboxes
deployed via AdminAPI. This is done by changing the sandbox code to
call wrapper scripts that call mysqld inside a loop instead of calling
mysqld directly.

Some tests needed to be changed since they were not expecting the
restart functionality to succeed.

Change-Id: Ibdb43bf507e3fcdc3374fab666a48383ed4895fa