BUG#29707077 \WATCH ONLY HANDLES CTRL+C AT REFRESH
Interrupting the \watch command via CTRL-C did not work on Windows,
users had to wait till next screen refresh.

Problem was caused by the implementation of the internal sleep function
which was used to wait before refreshing the screen between subsequent
executions of a report, it could not be properly interrupted. This is
now fixed, other Shell functionalities which use the same function are
correctly detecting the CTRL-C key combination.

Additionally, queries which are executed through the session object
received by the report function are now automatically cancelled when
CTRL-C is pressed. This change affects both \show and \watch commands.

Change-Id: Ibcd8b42934f5e02e9c5a3cf68b549d65e2ec6fcb