WL#12446 DevAPI: Support connection attributes
This patch introduces the handling of user defined connection attributes for
both classic and X protocols.

The connection-attributes option is introduced as a connection option which
accepts the following values:

- Booleans: true/false/1/0
- An array of attribute=value pairs
- A dictionary with the attribute/value pairs (Not supported in URI)

Default behavior is that a predefined set of connection attributes will be
stored when a session is created: _client_name, _client_version, program_name,
etc.

When 0 or false is used, the behavior depends on the protocol:

- Classic: only the attributes defined by the client lib are registered
  (program_name is excluded).
- X: no attributes are registered.

When additional attribute/value pairs are defined, behavior depends on whether
the target server supports connection attributes:

- For classic sessions there are no restrictions, the additional attributes will
  be registered.
- For X, if the server supports connection attributes they will be registered
  otherwise an error will be raised.

Connection Options Refactoring

The connection options may either receive a single value, an array of values
or in some cases both types of values with different behaviors. For this reason
a refactoring was done on the Connection_options to support:

- Rejecting arrays of values on attributes that must receive a single value.
- Enable handling of attributes with multiple values.
- Enable handling of connection-attributes.

After this patch the Connection_attributes class no longer inherits from
Nullable_options.

URI Parser Update

The URI parser was updated to properly handle attributes with single and array
values when filling the Connection_options object.

Additional Changes
- shell.dumpRows was updated to return the number of printed records.
- Api_connection tests validation files were unified: in the past we did not
  have the conditional output verification options we have now, so a feature to
  allow defining custom validation file was introduced. Continuing with the
  multiple validation file support is way more complex to maintain so the
  feature was removed from the shell script tester.

Change-Id: If2f0798ecdb04bf1257a7ed47f15ce196da63bb1