BUG#29809560 : addInstance() does not validate if server_id is unique
The <Cluster>.addInstance() operation does not validate if the server_id
of the joining instance is unique among all cluster members. Although,
the use of instances with unique server_id is not mandatory for a Group
Replication (GR) group to work properly (because all replication
channels internal used by GR use --replicate-same-server-id=ON), it is
recommended that all instances in a replication stream have a unique
server_id.

This will avoid issues when connecting more slaves to the
clusters using a classic asynchronous replication channel (for read
scale-out), or when connecting two clusters with a asynchronous channel
(for disaster recovery). Furthermore, the auto_increment settings
updated by AdminAPI operations implicitly assume that the server IDs of
all cluster members are unique. Thus, this recommendation should be
"enforced" when adding a new instance to cluster.

This patch fixes this issue by adding a new validation to the
<Cluster>.addInstance() operation, aborting the operation with an error
if the server_id of the joining instance is already being used by
another member of the cluster.

A test case was added for regression.

Change-Id: I48a11b7e228523b53ba938d423a305207de9f919