BUG#29451154 Shell Fails To Output Stored Procedure Results In Loops
This patch corrects the handling of multiple results produced by a single
statement: a stored procedure.

The behavior was wrong in both X and MYSQL protocols with different reasons:

- X Protocol: was not refreshing the metadata as a new result was picked, this
  caused the data loading to be aborted because of mismatched data types between
  the held metadata (from the very first result) and the data contained on the
  subsequent results.
- MySQL Protocol: the handling of multiple results was completely broken in the
  sense that the next result was not being read at all.

In addition, the Result dumper was updated to print a blank line between results
when multiple are found for the same statement.

Change-Id: I9dc634f21f5d0f8844902dc8ebffd65ff37e2f05