BUG#29634828 ADMINAPI SHOULD HANDLE LOCALHOST AND SANDBOXES BETTER
On Debian-based hosts, hostname by default resolves to 127.0.1.1 IP
address, which does not match a real network interface. This is not
supported by the Group Replication communication layer, making sandboxes
deployed on such hosts unusable, unless a manual change to the
configuration file is made.

The configuration file of the sandboxes created by Shell has been
updated, it now contains an additional line:
  report_host = 127.0.0.1
This will allow for the sandboxes to be used on Debian-based hosts
without any additional manual changes. Since sandbox instances are meant
to be used locally and for testing purposes only, the `report_host`
variable is set to the loopback IP address.

For more information visit:
https://dev.mysql.com/doc/refman/en/replication-options-slave.html#option_mysqld_report-host

Change-Id: Ia35fdc57301f8ca1b7970fa5610d6c24a039b817