BUG#29271400 createCluster() should not be allowed in instance with metadata
The dba.createCluster() was allowed in a target instance which already
had a populated Metadata schema, when instance was in that Metadata.
The Metadata present in the target instance was overridden without the
consent of the end-user, which was unexpected.

The dba.createCluster() will now throw an exception if it is called in
an instance which has the Metadata schema and it belongs to that
Metadata and will suggest to either drop the schema or reboot the
cluster.

Change-Id: I6eeac91b6a55a033fbd9b0965ed55c59d3dd963d