BUG#28727505 BUG#29765093 : extend instance check to persisted (only) values
The dba.checkInstanceConfiguration() and dba.configureInstance() operations
were not checking the validity of persisted configurations, which can be
different from the corresponding system variable value, in particular when
changed with SET PERSIST_ONLY. This can lead those operations to report
wrong or inaccurate results, for example: report that the instance
configuration is correct when in reality the persisted configuration is
invalid and wrong settings will be applied after a restart of the server
(BUG#29765093), or inaccurately report that a server update is needed when
only a restart is required (BUG#28727505).

This patch fixes the above issue by validating the persisted values for the
required configuration, when available and different from the current system
variable value. Some small output/behavior changes where required to fix the
issue, namely:
- the previous "restart" action that appeared in the JSON result of the
  dba.checkInstanceConfiguration() was replaced by "server_update+restart",
  which was not used, maintaining the same meaning "Update read-only variable
  and restart the server" in the displayed table with the needed
  configuration changes;
- the "restart" action is now used to simply report a "Restart the server",
  without any required variable update.

Tests were added for regression and existing tests updated to match the
new change of behavior.

Change-Id: I301be230b3e0bbb4705efd22d5bc1743998f44c7