WL#13208 BUG#29630591 BUG#29790569 BUG#29823941 BUG#29825836 BUG#29878581 BUG#29901010 BUG#29901784 BUG#29901884 BUG#29901902 BUG#29901967 BUG#29905066 BUG#29905246 AdminAPI: Automatic handling of Clone GR integration
This patch introduces the AdminAPI automatic handling of clone GR
integration, in other words, it introduces automatic node provisioning
by making use of the Clone plugin, introduced in 8.0.17.

The following high-level features/changes are introduces:

dba.checkInstanceState():
  - improved to obtain an accurate state GTID state of the target instance in
    relation to the cluster.

dba.checkInstanceConfiguration():
  - extended to verify if clone is available but disabled on the target
    instance.

dba.createCluster():
  - installs the clone plugin by default when available.
  - extended with two new options: disableClone and gtidSetIsComplete.

<Cluster>.setOption():
  - extended to support disabling clone in a cluster.

<Cluster>.options():
  - extended to display the value of disableClone of the target cluster

<Cluster>.addInstance():
  - extended with a new option: recoveryMethod to allow the user to explicitly
    determine which recovery method must be used: auto, clone or regular
  - extended with checks to determine which type of recovery must be used and
    prompt the user accordingly.
  - extended with a new option: waitRecovery to indicate if the command shall
    wait for the recovery process to terminate or not and its verbosity level.

<Cluster>.status():
  -  extended to to provide information about an on-going recovery process of
     any cluster members.

Handling of recovery accounts extended to include the required grants
to use Clone and also to overcome the limitations seen when clone is
used.

BUG#29630591: CLUSTER.CHECKINSTANCESTATE() INCORRECT OUTPUT WHEN BINARY LOGS WERE PURGED

<Cluster>.checkInstanceState() lacked the ability to check if the binary
logs have been purged from all cluster members, resulting in erroneous
output values. This patch extends checkInstanceState() to be able to
validate the value of GTID_PURGED on all cluster members and provide the
correct output and also an informative message mentioning the possible
actions to be taken.

BUG#29790569: ADDINSTANCE/REJOININSTANCE NOT CHECKING THE GTID STATE IN REGARDS TO THE CLUSTER

<Cluster>.addInstance()/rejoinInstance() did not make use of the checks
performed by <Cluster.>checkInstanceState() in order to verify the GTID
status of the target instance in relation to the cluster. On the event
of all cluster members having the binlogs purged, the addInstance()
command would succeed but the instance would never be able to join the
cluster as distributed recovery fails to execute. This patch extends
both commands to make use of the checks performed by
<Cluster>.checkInstanceState() and provide informative error messages.

Change-Id: If6d4aab8301fb612b7387927974c29ff2c6e2a40