BUG#29754915 : rejoininstance fails if cluster members are recovering
For MySQL servers with version 8.0.16, the <Cluster>.rejoinInstance()
operation fails when one or more cluster members are in RECOVERING state
because Group Replication communication protocol cannot be obtained.
More specifically, the group_replication_get_communication_protocol()
User-Defined function (UDF) fails because it can only be executed if all
members are ONLINE.

This patches fixes the issue when rejoining an instance by printing a NOTE to
notify the user and proceeding with the execution of the operation in case
the group_replication_get_communication_protocol() UDF fails with MySQL error
1123 (ER_CANT_INITIALIZE_UDF).

NOTE: Starting from MySQL 8.0.17, the
group_replication_get_communication_protocol() UDF no longer issues an error
if some member is RECOVERING.

Tests were added for regression.

Change-Id: Ib42514f8a8d300745ddfb7dbf798f2046426348f