BUG#25025880 BUG#25253864 BUG#28542803 BUG#25740356 Misc Help Bugfixes
Bug #25025880 GLOBAL SYS OBJECT IN JS MODE UNDOCUMENTED

The global sys object provided for JavaScript was not being shown either on
the shell help nor the API docs, this patch adds the help information to both
places.

BUG#25253864 UNDOCUMENTED FUNCTION DIR() - REMOVE OR DOCUMENT

The global dir() function provided for JavaScript was undocumented, this patch
fixes this for both the help system and the API docs.

BUG#28542803 MANUAL REHASH REQUIRED AFTER CREATE TABLE ON ANOTHER SESSION

The support for tables and collections as properties on a Schema object has a
limited scope by design. This bug is being closed by clearly documenting the
conditions required for a table or a collection to be available as a property
in a schema object.

BUG#25740356 COLLECTION ADD HELP A BIT CONFUSING ON MYSQLX.EXPR - OR WRONG, OR NOT WORKING?

Usage of expressions to add a document was documented as mysqlx.expr(<JSON String>)
which is incorrect as JSON string includes objects, arrays, numbers, etc. The help
has been updated to properly indicate that it the parameter should be a string
representing a JSON object. Examples were also added to clarify even more.

Change-Id: I18a13a11b951e4604de1c01f27b207cece83f62f