BUG#29996051 Removing Interactive Wrappers
The interactive wrappers were initially introduced to have a clean separation
between the interactive logic and the API logic.

The shell code has evolved in a way that such separation is no longer required
and also having the wrappers has proven to complicate and even cause code
duplication on the functions that require user interaction.

This patch integrates the interactive bits into the API classes itself, removing
the need of the interactive laywer.

The following classes are being removed:
- Interactive_object_wrapper
- Interactive_dba_cluster
- Global_dba
- Global_shell

Change-Id: I6b122835943dd95365081829b5a5cec8225ace3c