WL#12773 BUG#29629121 BUG#29867008 BUG#29559303 BUG#29868204 BUG#29869023 BUG#29862474 Simplification of internal recovery accounts
This simplifies the number of recovery accounts that are created.
After this patch a only a single recovery account is created for each
member of the cluster, even if the ipWhitelist option is used.

Furthermore, the recovery user with the "localhost" host part is no
longer created (only '%'). This can lead to an issue in local
experimental setups using "localhost" for the report_host. It is a
known limitation, but not considered a problem because "localhost" is
not expected to be used (especially not in a real production
settings).

This patch also deprecates the superReadOnly option for the
createCluster and rebootClusterFromCompleteOutage commands. The
super-read-only mode is now automatically disabled.

Existing tests were modified and updated according to the new number
and name format of the recovery accounts. No new tests were added because
existing tests already covered the test plan proposed by QA.

BUG#29629121: track created users, do not use pattern to guess when
cleaning up

This bug was fixed by using the information about the recovery users
stored in the metadata or in the mysql.slave_master_info table when
dropping the recovery user. Users are no longer dropped according to a
regular expression.

BUG#29867008: removeinstance(), rejoininstance() shouldn't drop all accounts

This bug was only partially fixed in the context of the WL.
This wl fixed it by dropping the recovery account that belongs to
the instance removed instance at the primary instance and then waiting
for that information to be replicated to the instance being removed
before actually removing the instance from the cluster.

BUG#29559303: cluster.dissolve() fails

This bug was fixed by handling both recovery account formats used by
the shell: the new name format defined by this worklog
'mysql_innodb_cluster_<server_id>' and the old account format
'mysql_innodb_cluster_r<random_number>' when dropping a recovery user
of an instance. This is done in both cluster.removeInstance and
cluster.dissolve.

BUG#29868204: cluster.addinstance() fails complaining about read-only
mode

This bug was a symptom of a new check that was added during the
context of this worklog but has since been removed. The check needed
to create a new user and that check was being done before the SRO was
disabled by the addInstance operation. Since the check is no longer made,
nothing really needed to be done to fix this issue.

BUG#29869023: cluster.rejoininstance() fails: instances cannot
communicate with each other

This bug was a symptom of a new check that was added during the
context of this worklog but has since been removed. The check executed
a "change master" that was not correctly handling cases where the
servers were using ssl. This issue has since been fixed even tough the
check is not currently used.

BUG#29862474: dba.checkinstanceconfiguration fails when
--super-read-only is on

This bug was a symptom of a new check that was added to
dba.checkInstanceConfiguration() during the context of this worklog
but has since been removed. The new check would not work if the target
instance had SRO enabled. This was a new requrement, as previously
none of the checks made by dba.checkInstanceConfiguration() would
require SRO to be disabled and that was the cause of the issue. Since
the check has since been removed, there was nothing that needed to be
done to fix this issue.

Change-Id: Ib567e3f2c7a3bd23c515b04d91354fb1e6352cf9