BUG#29897911 UNABLE TO DEFINE A COMPLEX PYTHON PLUGIN: 'IMPORT' DOES NOT WORK
This patch fixes the plugin load logic so it:
- Properly adds the plugin path to the sys.path
- Properly defines the plugin __file__ variable

Plugin Group Support

Now, a plugin folder may contain either:
- A single plugin definition
- Multiple plugin definitions

This is determined by the presence of the initialization file, if it is
present, the plugin is assumed to be a single plugin. If it is not present
the plugin is assumed to be a plugin group, this is, it may contain the
definition of multiple plugins.

A plugin group is defined by the main plugin folder, containing sub-folders.
Each subfolder will be scanned for a standard plugin definition, and loaded
following the standard plugin load logic.

Only the main plugin folders support multiple plugin definitions, this is,
this logic is not recursive.

Change-Id: I06d550bb323db07ec3c1d0c641ed9bdc143bd021