BUG#25503159 : addInstance() does not drop recovery user when it fails
The <Cluster>.addInstance() operation creates internal recovery users
which are required by the Group Replication (GR) recovery process.
However, when the operation fails, namely if GR fails to start, the created
recovery users are left in the database. The created recovery user should be
removed in case the operation fails.

This patch fixes this issue by dropping the created recovery user in the
primary instance in case the <Cluster>.addInstance() fails.

A new test was added for regression.

Change-Id: I7b5e47c33e6550dc4f5ac180b8776b6035df0d62