BUG#29965071 : UNABLE TO ADD AN 8.0 INSTANCE TO A 5.7 INNODB CLUSTER
Whenever adding an 8.0.17 instance to a cluster consisting of 5.7
instances, the <Cluster>.addInstance() command would fail with an error
caused by a SQL sintax error originating from the query to grant the
required grants for the recovery account. Specifically, this query would
use the BACKUP_ADMIN grant to the newly created recovery user which is
created on the primary instance of the cluster that is a 5.7 instance
which does not support BACKUP_ADMIN.

This patch fixes the issue by ensuring that BACKUP_ADMIN is only granted
to the recovery accounts when ALL cluster members support it and also
the target instance supports it.

On top of that, whenever changing the value of disableClone to false
with the <Cluster>.setOption() command, the AdminAPI now ensures that
the BACKUP_ADMIN is added to all recovery accounts in use by the
cluster.

Change-Id: Id1d90b32ccc9057a03527f69d78c2e468d5639a5