WL#12763 Add \edit (\e) and \system (\!) Shell Commands
This WL introduces two new Shell commands:

\edit (\e): Allows editing the command to be executed in the default
system editor.
The editor is chosen using the EDITOR and VISUAL environment variables
(in that order). If these variables are not set, Shell will use notepad.exe
on Windows and vi on all remaining platforms. Shell creates a temporary
file with the command to be edited, then launches an editor with that
file. Once the editor is closed, the current contents of the temporary
file is presented to the user as the next command to be executed. If no
arguments are given, the file will contain the command executed prior to
the \edit command. If history is empty, file will be empty as well. If
arguments are provided, they are going to be used as an initial contents
of the file.
It is also possible to invoke this command by pressing CTRL-X CTRL-E.

\system (\!): Allows executing a system command, displays the results in
the Shell.
It executes the command given as arguments, if no arguments are provided
it displays help on \system. If a pager is active, it's not going to
receive the output of the executed system command.

Change-Id: Ib28ab0e923bdc20c303403aba4543281bfa94e73