BUG#25267603 : addInstance() tries to use offline node after lost quorum
In a lost quorum scenario with the majority of the cluster offline
except the primary, after reestablishing the quorum and adding a new
instance to the cluster, it was not possible to remove and add the
previous primary instance to the cluster because the operation failed
trying to contact offline instances.

The operation was failing because the feature to verify if a Group
Replication (GR) protocol upgrade is required was not considering the
possibility of some cluster instances being offline (not reachable).

This patch fixes this issue by improving the GR protocol upgrade
handling for the <Cluster>.removeInstance() operation, attempting to
connect to other cluster members and using the first reachable member
for this purpose.

A new test was added for regression.

Change-Id: I6fb7bc285943aa58e4c9d4f065e50ad816607c8b