WL#13294 : option to log AdminAPI SQL
This patch provides a new feature to allow SQL statements executed by the
AdminAPI operations to be logged by the MySQL Shell. The SQL statements
are logged with the INFO log level and a new `dba.logSql` shell option
was added to control the SQL statement logging, accepting the following
values:
 - 0: no SQL statement logging (default);
 - 1: log SQL statements, except SELECT and SHOW;
 - 2: log all SQL statements;

A corresponding '--dba-log-sql' command line option was also added to set
a custom value for `dba.logSql` at startup.

Additional small fixes were also included in this patch, namely:
- the validation of the allowed range of values for the `verbose` shell
  option was fixed;
- the determination if single-primary mode was used for the
  mysqlshdk::gr::get_members() function was improved/fixed;

NOTE: It is possible to display the logged SQL statements in the console
using the already existing feature to show logged INFO messages by enabling
verbose (setting it with a value > 0).

New tests were added for regression.

Change-Id: I4c951ca30f93c0f4c19b0f210abdff1c7e00144c