Adding support for tests with multiple versions of the server.
run_unit_tests update
=====================
The run_unit_test application now can work with tests using multiple
versions of the server, the requirements include:

- Passing to --record, --replay, --direct the different versions to be
  used in the format of:
  <version>[+<version>]*

  The first server is the same as the base server used for tests, while
  the rest are the additional servers to be available. For this to work
  the path to the mysqld binaries must be defined in environment variables
  as follows:

  If 2 versions are provided and they differ in the major or minor version
  the path to the corresponding mysqld must be provided in an environment
  variable named: MYSQLD<major><minor>_PATH

  If the condition above is not met, then the path to the binary of every
  additional server must be provided in individual environment variables
  named: MYSQLD<majot><minor><patch>_PATH

Testutils Update
================
The testutils.deploy*Sandbox signature changed as follows:
- testutils.deployRawSandbox(port[, cnfOptions[, deployOptions]]
- testutils.deploySandbox(port[, cnfOptions[, deployOptions]])

The supported deployOptions include:
- createRemoteRoot: bool
- mysqldPath: string

The environment variables for the extra servers will be available as normal
variables on scripting, they can be used to deploy a sandbox of a version
different than the base server version used in the tests.

Paratrace Update
================
The paratrace will automatically pass the additional servers to the RUT call
letting as base server the base server for recording and as additional any
other server added with the add function.

Writing Tests
=============
The environment variables used to register the paths to the mysqld binaries
are made available as normal variables on the scripts located at:

- js_mixed_versions
- py_mixed_versions

These tests MUST be executed ONLY if the required variables are defined.
For this purpose a macro DEF(var_name) has been added, which will return
true if the variable is defined, and false if the variable is not defined.

Example:

Running RUT using a secondary server, i.e. 5.7.27 is done by passing
8.0.16+5.7.27 to the record/replay/direct parameter.

8.0.16 Will be the base server version.
5.7.27 Will be the secondary server version.

RUT will validate that an environment variable MYSQLD57_PATH is defined with
the path to the 5.7.27 server binary.

MYSQLD57_PATH will be a variable available on the scripts and could be used
to enable the script if defined, as follows:

//@{DEF(MYSQLD57_PATH)}
  - Enables a JS test script if the MYSQLD57_PATH variable is defined.

Or to pass the mysqldPath to the sandbox deployment functions.

Change-Id: I8778619c9b7312274835533ce30f91b6dee8fd8b